all: convert tests for flowcontrol
